requirements.txt
python-dotenv
openai
tiktoken
beautifulsoup4
pandas
pypdf
xlsxwriter
supabase
pymupdf
duckduckgo_search
fastapi
uvicorn
PyJWT
youtube-transcript-api
numpy
stripe
pytest-playwright
modal
nltk
trafilatura
markdown
pyyaml
datasketch